improve: nlp tool calling
Improvements include better control over Tool Calls. Upon completion of inference, the final structure of the generated Human+Tool Calls+Tools messages is validated to be correct.
All messages generated in the inference process for a user are temporarily stored in a memory buffer and do not impact history storage until the inference is complete and they are validated.
This improves response times and the stability of the assistant.

A possible bug was also fixed due to which the stream+sentence mode was not being fully exploited.